test+lint: suites de contratos, puertos y rendimiento; mypy estricto y ruff limpios
- tests nuevos: errors, observability, runtime, processes, cli, ports, agent lifecycle,
  perf semantics, notrack retry, diagnostic selftest (317 pruebas, 82% cobertura)
- pyproject: version dinámica desde autonoma._version, ruff con selección documentada,
  mypy strict (overrides para dependencias opcionales), coverage fail_under=80
- mypy --strict sin errores; ruff check sin avisos
- fixes de producción hallados por los tests:
  * PanicController.unregister_cleanup comparaba por identidad: los bound methods
    nunca coincidían y las limpiezas se acumulaban en cada /key (fuga real)
  * DiagnosticReport.crash al serializar data_root=Path; normalizado con _root_payload
  * métricas de turno dobles (turn.completed nunca existía) -> turn.ok/cancelled/failed
  * BOM filtrado en stdout con utf-8-sig; timeout inválido es ConfigurationError, no timeout
  * chat() encadena la causa de red (log) sin filtrar el detalle al modelo
  * search_engine: errores tipados (ProviderHttpError/Contract/SearchBackend) en lugar de RuntimeError
  * tool_schemas() copia profunda; schemas_payload() tuple congelado para la ruta caliente
  * fs.close()/Agent.close() conforman ResourcePort; trace_id fijado al emitir el log

+++ autonoma-agent/autonoma/observability.py
@@ -147,7 +147,7 @@ class DurationSummary:
 def _percentile(ordered: Sequence[float], fraction: float) -> float:
     if not ordered:
         return 0.0
-    index = min(len(ordered) - 1, max(0, int(round(fraction * (len(ordered) - 1)))))
+    index = min(len(ordered) - 1, max(0, round(fraction * (len(ordered) - 1))))
     return ordered[index]
 
 
@@ -324,7 +324,7 @@ class _TraceIdFilter(logging.Filter):
 
     def filter(self, record: logging.LogRecord) -> bool:
         if not hasattr(record, "autonoma_trace_id"):
-            record.autonoma_trace_id = current_trace_id()  # type: ignore[attr-defined]
+            record.autonoma_trace_id = current_trace_id()
         return True
